generalized extractionts to extractionTabular
