+++ renderer/components/mainPages/extraction_ts.jsx
@@ -1,4 +1,4 @@
-import ExtractionTSCanvas from "../extraction_ts/extractionTSCanvas"
+import ExtractionTabularData from "../extraction_ts/extractionTabularData"
 import React from "react"
 import ModulePage from "./moduleBasics/modulePage"
 
@@ -6,7 +6,7 @@ const ExtractionTSPage = ({ pageId, configPath = "" }) => {
   return (
     <>
       <ModulePage pageId={pageId} configPath={configPath}>
-        <ExtractionTSCanvas />
+        <ExtractionTabularData />
       </ModulePage>
     </>
   )
