+++ renderer/styles/globals.css
@@ -129,6 +129,10 @@ a {
   margin-top: 15px;
 }
 
+.margin-top-30 {
+  margin-top: 30px;
+}
+
 .margin-left-10 {
   margin-left: 10px;
 }
